useErrorRouting is now useNetworkStatus, and app-error page is now a 500 page
